SimpleLink lpf3-8.40.00.61 update
This commit updates to the SimpleLink F3 8.40.00.61 SDK release.

+++ source/ti/ble5stack_flash/hal/src/common/hal_drivers.c
@@ -178,7 +178,7 @@ void HalDriverInit (void)
  *
  * @return  None
  **************************************************************************************************/
-uint16 Hal_ProcessEvent( uint8 task_id, uint16 events )
+uint32 Hal_ProcessEvent( uint8 task_id, uint32 events )
 {
   uint8 *msgPtr;
