+++ source/ti/ble5stack_flash/icall/inc/ble_dispatch_lite.h
@@ -91,7 +91,7 @@ extern void ble_dispatch_liteInit(uint8_t taskId);
 /*
  * Task Event Processor for the task
  */
-extern uint16 ble_dispatch_liteProcess(uint8_t taskId, uint16_t events);
+extern uint32 ble_dispatch_liteProcess(uint8_t taskId, uint32 events);
 
 /*
  * BLE direct API message parser
